rendered props to bypass refresh issue
