+++ Frontend/trivializer/src/components/EditGameView.js
@@ -1,5 +1,9 @@
 import React, { Component } from "react";
 
+/**
+ * EditGameView
+ * - view to edit selected game and handle updates
+ */
 class EditGameView extends Component {
     constructor(props) {
         super(props);
